sampleprj/requirements.txt
appdirs==1.4.0
dj-database-url==0.4.2
Django==1.10.5
django-crudbuilder==0.2.5
django-extensions==1.7.6
django-tables2==1.3.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
six==1.10.0
